Basic improvements that were needed to bootstrap a development environment
This is basically the state of my working tree with changes to create a
docker-based environment. I added basic port-forwarding and updated some
details of the docs as well.

In the end I decided to rewrite `initial-setup.sh` to make it less
error-prone (we shouldn't ignore failed CouchDB calls silently and
provide a better help message).